Removed Watchlist Service.  Moved TransactionError to ResourceTransactable.
